code implemented
Can Merklefy and reverse Merklefy an io.Reader, io.ReadSeeker
respectively.
Where the reverse Merkle tree is basically a file/io.ReadSeeker
read in reverse blocks and not necessarily with the bytes
exactly reversed -- Work in progress though.
The purpose of the reverse tree is to be able to quickly figure out
what differences occur in a tree when data is concatenated
to the front, middle or end.

For example:
* Tradition single Merkle Tree:
+ | A | B | C |
[Root]
	[Root] --> h(h(h(A)+h(B)) + h(C))
	/     \
|h(h(A)+h(B))|-|h(C)|
      /		\
  |  A  |  B  |  C  |

With data added to
a) Front:
	   [Root] --> h(h(h(X)+h(A)) + h(h(B)+h(C)))
	 /        \
	/          \
| h(h(X)+h(A)) |---| h(h(B)+h(C)) |
      /		     \
  | X	| A   | B   | C	  |

b) Middle:
	   [Root] --> h(h(h(A)+h(B)) + h(h(X)+h(C)))
	 /        \
	/          \
| h(h(A)+h(B)) |---| h(h(C)+h(X)) |
      /		     \
  | A	| B   | X   | C	  |

c End:
	   [Root] --> h(h(h(A)+h(B)) + h(h(C)+h(X)))
	 /        \
	/          \
| h(h(A)+h(B)) |---| h(h(C)+h(X)) |
      /		     \
  | A	| B   | C   | X	  |

will render the entire tree's blocks as undifferable since
the blocksize spills over to new blocks.
Still a work in progress though for the reverse Tree.